dermascan/model/requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
numpy==1.26.4
pandas==2.0.2
Pillow==9.5.0
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
tqdm==4.65.0